Seperate echo process and wps_testing
