+++ .github/labeler-files.yml
@@ -32,7 +32,7 @@ ci/doc:
 ci/tests:
   - tests/**/*
   - weaver/processes/wps_default.py
-  - weaver/processes/wps_testing.py
+  - weaver/processes/echo_process.py
 
 ## Triage
